# Improve performance of normalize_vector